+++ bungee.yml
@@ -1,4 +1,4 @@
 name: SkinsRestorer
 main: skinsrestorer.bungee.SkinsRestorer
-version: 10.9
+version: 11.0
 author: Th3Tr0LLeR
